+++ orc8r/cloud/go/services/accessd/accessd/main.go
@@ -23,7 +23,7 @@ import (
 	"magma/orc8r/cloud/go/service"
 	"magma/orc8r/cloud/go/services/accessd"
 	"magma/orc8r/cloud/go/services/accessd/protos"
-	"magma/orc8r/cloud/go/services/accessd/servicers"
+	servicers "magma/orc8r/cloud/go/services/accessd/servicers/protected"
 	"magma/orc8r/cloud/go/services/accessd/storage"
 	"magma/orc8r/cloud/go/sqorc"
 	storage2 "magma/orc8r/cloud/go/storage"
